pyproject.toml
[tool.poetry]
name = "dolar-api"
version = "0.1.0"
description = "Argentina Dolar API"
authors = ["Osvaldo Demo <demoosvaldo@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.4"
django-rest-framework = "^0.1.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
